Split rust_admin config.rs validation and CORS helpers into submodules
config.rs keeps the Config struct and from_env constructor; auth-value
validation and cookie SameSite parsing move to config/validation.rs,
CORS layer construction to config/cors.rs, tests to config/tests.rs.
Also drops a local non_empty_env duplicate in favor of the
autvid_common helper.
