+++ crates/rust_stream/src/tests.rs
@@ -1,4 +1,3 @@
-
 #![allow(clippy::unwrap_used)]
 use std::{
     sync::{
